Refactor permission handling.
Projects are now associated with classes so only a single setting file is
required for each class as they are shared by all projects within a class.

There is now a `users.json` which contains all the users in the system and
their allowed password.

TODO: Project specification in the class file should specify what plugins are
used on a given project.